backend-drf/requirements.txt
django==6.0.1
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
django-cors-headers==4.9.0
python-decouple==3.8
joblib==1.5.3
scikit-learn==1.7.2
nltk==3.9.2
